+++ src/main/java/com/webdev/cheeper/repository/RoomRepository.java
@@ -19,17 +19,17 @@ public class RoomRepository extends BaseRepository {
             
             ResultSet rs = stmt.getGeneratedKeys();
             if (rs.next()) {
-                room.setId(rs.getLong(1));
+                room.setId(rs.getInt(1));
             }
         } catch (SQLException e) {
             e.printStackTrace();
         }
     }
 
-    public Optional<Room> findById(Long id) {
+    public Optional<Room> findById(Integer id) {
         String sql = "SELECT * FROM room WHERE id = ?";
         try (PreparedStatement stmt = db.prepareStatement(sql)) {
-            stmt.setLong(1, id);
+            stmt.setInt(1, id);
             ResultSet rs = stmt.executeQuery();
             if (rs.next()) {
                 return Optional.of(mapResultSetToRoom(rs));
@@ -40,14 +40,14 @@ public class RoomRepository extends BaseRepository {
         return Optional.empty();
     }
 
-    public List<Room> findByUserId(Long userId) {
+    public List<Room> findByUserId(Integer userId) {
         String sql = "SELECT r.* FROM room r " +
                     "JOIN room_participants rp ON r.id = rp.room_id " +
                     "WHERE rp.user_id = ? " +
                     "ORDER BY r.created_at DESC";
         List<Room> rooms = new ArrayList<>();
         try (PreparedStatement stmt = db.prepareStatement(sql)) {
-            stmt.setLong(1, userId);
+            stmt.setInt(1, userId);
             ResultSet rs = stmt.executeQuery();
             while (rs.next()) {
                 rooms.add(mapResultSetToRoom(rs));
@@ -58,7 +58,7 @@ public class RoomRepository extends BaseRepository {
         return rooms;
     }
 
-    public Optional<Room> findPrivateRoomBetweenUsers(Long user1Id, Long user2Id) {
+    public Optional<Room> findPrivateRoomBetweenUsers(Integer user1Id, Integer user2Id) {
         String sql = "SELECT r.* FROM room r " +
                     "JOIN room_participants rp1 ON r.id = rp1.room_id " +
                     "JOIN room_participants rp2 ON r.id = rp2.room_id " +
@@ -67,8 +67,8 @@ public class RoomRepository extends BaseRepository {
                     "AND rp2.user_id = ? " +
                     "LIMIT 1";
         try (PreparedStatement stmt = db.prepareStatement(sql)) {
-            stmt.setLong(1, user1Id);
-            stmt.setLong(2, user2Id);
+            stmt.setInt(1, user1Id);
+            stmt.setInt(2, user2Id);
             ResultSet rs = stmt.executeQuery();
             if (rs.next()) {
                 return Optional.of(mapResultSetToRoom(rs));
@@ -79,11 +79,11 @@ public class RoomRepository extends BaseRepository {
         return Optional.empty();
     }
 
-    public void addParticipant(Long roomId, Long userId) {
+    public void addParticipant(Integer roomId, Integer userId) {
         String sql = "INSERT INTO room_participants (room_id, user_id) VALUES (?, ?)";
         try (PreparedStatement stmt = db.prepareStatement(sql)) {
-            stmt.setLong(1, roomId);
-            stmt.setLong(2, userId);
+            stmt.setInt(1, roomId);
+            stmt.setInt(2, userId);
             stmt.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -92,7 +92,7 @@ public class RoomRepository extends BaseRepository {
 
     private Room mapResultSetToRoom(ResultSet rs) throws SQLException {
         Room room = new Room();
-        room.setId(rs.getLong("id"));
+        room.setId(rs.getInt("id"));
         room.setName(rs.getString("name"));
         room.setPrivate(rs.getBoolean("is_private"));
         room.setCreatedAt(rs.getTimestamp("created_at"));
